Split command palette reducers by responsibility
